migrate some more bool arrays to bit arrays

See merge request graphviz/graphviz!2356

+++ lib/neatogen/sgd.h
@@ -14,10 +14,10 @@ typedef struct term_sgd {
 
 typedef struct graph_sgd {
     size_t n; // number of nodes
-    int *sources; // index of first edge in *targets for each node (length n+1)
+    size_t *sources; // index of first edge in *targets for each node (length n+1)
     bitarray_t pinneds; // whether a node is fixed or not
 
-    int *targets; // index of targets for each node (length sources[n])
+    size_t *targets; // index of targets for each node (length sources[n])
     float *weights; // weights of edges (length sources[n])
 } graph_sgd;
